+++ packages/tour-kit/src/components/tour-kit-overlay.tsx
@@ -7,14 +7,14 @@ interface Props {
 	visible: boolean;
 }
 
-const Overlay: React.FunctionComponent< Props > = ( { visible } ) => {
+const TourKitOverlay: React.FunctionComponent< Props > = ( { visible } ) => {
 	return (
 		<div
-			className={ classnames( 'tour-kit__overlay', {
+			className={ classnames( 'tour-kit-overlay', {
 				'--visible': visible,
 			} ) }
 		/>
 	);
 };
 
-export default Overlay;
+export default TourKitOverlay;
